requirements.txt
biopython>=1.81
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
scipy>=1.10.0
dendropy>=4.5.0
